third_party/WebKit/LayoutTests/html/details_summary/details-clone-expected.txt

Test to clone details and summary elements

PASS successfullyParsed is true

TEST COMPLETE
PASS targetMarkerPseudoId is clonedMarkerPseudoId
